指定した2つのコミットの差分を表示する

+++ src/kidivis/review.py
@@ -13,6 +13,8 @@ import urllib.parse
 import git
 import jinja2
 
+from . import diffimg
+
 LAYERS = ['.'.join(p) for p in product(['F', 'B'], ['Cu', 'Silkscreen', 'Mask'])] + ['Edge.Cuts']
 KICAD_CLI = '/mnt/c/Program Files/KiCad/9.0/bin/kicad-cli.exe'
 
@@ -37,6 +39,7 @@ def export_svgs(dst_dir_path, pcb_file_path):
             pcb_file_path = convert_path_to_windows(pcb_file_path)
 
     export_cmd = [KICAD_CLI, 'pcb', 'export', 'svg', '--mode-multi',
+                  '--fit-page-to-board', '--black-and-white',
                   '--layers', ','.join(LAYERS),
                   '--output', str(dst_dir_path), str(pcb_file_path)]
     res = subprocess.run(export_cmd)
@@ -45,6 +48,12 @@ def export_svgs(dst_dir_path, pcb_file_path):
         res.check_returncode()
 
 def extract_file(git_repo, commit_id, file_path, dst_path):
+    print(f'extracting file: commit={commit_id} file={file_path} dst={dst_path}')
+    if dst_path.exists():
+        return
+
+    dst_path.parent.mkdir(parents=True, exist_ok=True)
+
     if commit_id is None:
         # ワーキングツリーからファイル取得
         shutil.copy(file_path, dst_path)
@@ -81,30 +90,61 @@ class HTTPRequestHandler(http.server.BaseHTTPRequestHandler):
         parts = urllib.parse.urlparse(self.path)
 
         if parts.path == '/':
-            t = self.jinja_env.get_template('index.html')
-            svg_filename = make_svg_filename(self.pcb_path.name, 'F.Cu')
-            s = t.render(img_file_name=svg_filename).encode('utf-8')
-            self.send_response(200)
-            self.send_header('Content-Type', 'text/html')
-            self.send_header('Content-Length', len(s))
+            self.send_response(http.HTTPStatus.MOVED_PERMANENTLY)
+            self.send_header('Location', '/HEAD/WORK/F.Cu')
             self.end_headers()
-            self.wfile.write(s)
+            return
 
-        elif parts.path.startswith('/images/'):
-            file_name = parts.path[8:]
+        path = Path(parts.path)
 
-            path_fcu = self.tmp_dir_path / make_svg_filename(self.pcb_path.name, 'F.Cu')
-            if not path_fcu.exists():
-                export_svgs(self.tmp_dir_path, self.pcb_path)
-
-            with open(path_fcu, 'rb') as f:
-                svg_content = f.read()
-
-            self.send_response(200)
-            self.send_header('Content-Type', 'image/svg+xml')
-            self.send_header('Content-Length', len(svg_content))
+        if path.parts[0] != '/' or len(path.parts) != 4:
+            self.send_response(http.HTTPStatus.NOT_FOUND)
             self.end_headers()
-            self.wfile.write(svg_content)
+            return
+
+        diff_base, diff_target, layer = path.parts[1:]
+        if layer not in LAYERS:
+            self.send_response(http.HTTPStatus.NOT_FOUND)
+            self.end_headers()
+            return
+
+        base_dir = self.tmp_dir_path / diff_base
+        target_dir = self.tmp_dir_path / diff_target
+        pcb_filename = self.pcb_path.name
+
+        base_pcb_path = base_dir / pcb_filename
+        target_pcb_path = target_dir / pcb_filename
+        extract_file(self.git_repo,
+                     diff_base,
+                     self.pcb_path,
+                     base_pcb_path)
+        extract_file(self.git_repo,
+                     None if diff_target == 'WORK' else diff_target,
+                     self.pcb_path,
+                     target_pcb_path)
+
+        base_svg_path = base_dir / make_svg_filename(self.pcb_path.name, layer)
+        target_svg_path = target_dir / make_svg_filename(self.pcb_path.name, layer)
+
+        if not base_svg_path.exists():
+            export_svgs(base_dir, base_pcb_path)
+        if not target_svg_path.exists():
+            export_svgs(target_dir, target_pcb_path)
+
+        with open(base_svg_path) as f:
+            base_svg = f.read()
+        with open(target_svg_path) as f:
+            target_svg = f.read()
+
+        overlayed_svg = diffimg.overlay_two_svgs(base_svg, target_svg, True)
+
+        t = self.jinja_env.get_template('diffpcb.html')
+        s = t.render(svg=overlayed_svg).encode('utf-8')
+        self.send_response(200)
+        self.send_header('Content-Type', 'text/html')
+        self.send_header('Content-Length', len(s))
+        self.end_headers()
+        self.wfile.write(s)
 
 def handler_factory(*f_args, **f_kwargs):
     def create(*args, **kwargs):
@@ -124,7 +164,7 @@ def main():
     git_root = Path(git_repo.working_tree_dir)
     print(f'git work tree: {git_root}')
 
-    with tempfile.TemporaryDirectory() as td:
+    with tempfile.TemporaryDirectory(prefix='kidivis') as td:
         tmp_dir_path = Path(td)
         print(f'temporary directory: {tmp_dir_path}')
 
@@ -137,7 +177,11 @@ def main():
         with http.server.HTTPServer((args.host, args.port), create_handler) as server:
             print(f'Serving HTTP on {args.host} port {args.port}'
                   + f' (http://{args.host}:{args.port}/) ...')
-            server.serve_forever()
+            try:
+                server.serve_forever()
+            except KeyboardInterrupt:
+                input('press enter to exit...')
+                raise
 
     sys.exit(0)
